Add Pipeline#wrap to supersede implicit non-callable wrapping
Previously, any non-callable value passed to `Pipeline#pipe` was implicitly
wrapped in a lambda. There are a few reasons this behavior is not ideal:

1. It can lead to confusing scenarios when typos occurs (fat-thumbing a method
   as `calll` and wondering why everything is mysteriously broken)

2. Encourages use of magic behavior. If someone uses a non-callable as an
   argument, a read will have to ask themselves some questions:

   a) Is the argument a callable? This involves a potentially timely source-dive.

   b) If it's not a callable, I will likely have to source dive to see that
      `Piperator::Pipeline` automatically wraps a non-callable. It's not obvious
      that Pipeline should do this for the user. Principle of least surprise,
      yada-yada.

3. `respond_to?` adds a small percentage overhead which many people will pay
   for, and a small amount of people will benefit from.

On top of removing implicit wrapping, this commit adds `Pipeline#wrap`, which
is syntactic sugar for wrapping a value in a lambda. This way, we can favor
explicitness, but still provide an easy-to-use, non-ceremonious API.

So this example from the old API:

```ruby
   Piperator::Pipeline.pipe([1, 2, 3]).pipe(add1)
   # => [2, 3, 4]
```

Would now become:

```ruby
   Piperator::Pipeline.wrap([1, 2, 3]).pipe(add1)
   # => [2, 3, 4]
```